Add new class methods 'create' and 'add'.
Change the method 'add_and_put' into a private method '_add_and_put' which is used by the 'create' method.
The 'add' method is not implemented yet.

'create' vs 'add'
Create is used if a new icon is needed in the DB whereas add only increments an existing icon.
It is only a wrapper to simplfy the incrementation of a icon, the method 'incr()' could be used as well.